Delete emp

+++ src/main/resources/static/js/managerDashboard.js
@@ -10,6 +10,7 @@ export async function initManagerDashboard(){
     setupModal("readEmpModal", "all-emp-modal")
     setupModal("createShowModal", "create-show-modal");
     setupEmpForm()
+    setupEmployeeClick()
     setupShowForm()
     await loadEmployees()
 }
@@ -194,6 +195,7 @@ function setupEmployeeClick() {
             }
 
             const emp = await response.json();
+            console.log("Modtaget employee:", emp);
             displayEmployeeModal(emp);
 
         } catch (err) {
@@ -266,7 +268,7 @@ async function saveEmployee(id) {
     }
 
     alert("Medarbejder opdateret!");
-    document.getElementById("view-emp-modal").style.display = "none";
+    document.getElementById("emp-modal").style.display = "none";
     await loadEmployees();
 }
 
@@ -296,13 +298,23 @@ function setupShowForm(){
     };
 }
 // ------- DELETE EMP -------
-async function deleteEmp (employee){
+async function deleteEmp(employee) {
+    const id = employee.id;
+
+    const confirmed = confirm("Er du sikker på at du vil slette denne medarbejder?");
+    if (!confirmed) return;
+
     const response = await fetch(`http://localhost:8080/dashboard/manager/employees/${id}`, {
         method: "DELETE",
-        headers: {"Content-Type": "application/json"},
         credentials: "include"
     });
 
-
-
+    if (response.ok) {
+        alert("Medarbejder slettet");
+        document.getElementById("emp-modal").style.display = "none";
+        await loadEmployees();
+    } else {
+        const msg = await response.text();
+        alert("Fejl: " + msg);
+    }
 }
